Add finance item income

+++ app/src/main/java/com/example/coinkeeper/data/FinanceItemListRepositoryImpl.kt
@@ -15,7 +15,7 @@ object FinanceItemListRepositoryImpl: FinanceItemRepository {
 
     init {
         for (i in 0 until 10) {
-            val item = FinanceItem("Test$i","test$i", i.toDouble(),1)
+            val item = FinanceItem("Test$i","test$i", i,1)
             addItem(item)
         }
     }
